+++ testsuite/tests/polykinds/MonoidsTF.hs
@@ -14,6 +14,7 @@
 module Main where
 import Control.Monad (Monad(..), join, ap, liftM)
 import Data.Monoid (Monoid(..))
+import Data.Semigroup (Semigroup(..))
 
 -- First we define the type class Monoidy:
 
@@ -91,10 +92,13 @@ test2 = print (Sum 1 <+> Sum 2 <+> Sum 4)  -- Sum 7
 -- rather cumbersome in actual use. So, we can give traditional Monad and
 -- Monoid instances for instances of Monoidy:
 
-instance (MId (→) m ~ (), MComp (→) m ~ (,), Monoidy (→) m) 
+instance (MId (→) m ~ (), MComp (→) m ~ (,), Monoidy (→) m)
+       ⇒ Semigroup m where
+  (<>) = curry mjoin
+
+instance (MId (→) m ~ (), MComp (→) m ~ (,), Monoidy (→) m)
        ⇒ Monoid m where
   mempty = munit ()
-  mappend = curry mjoin
 
 instance Applicative Wrapper where
   pure  = return